requirements.txt
google-cloud-speech
python-dotenv